+++ backend/onyx/db/slack_channel_config.py
@@ -15,7 +15,7 @@ from onyx.db.models import SlackChannelConfig
 from onyx.db.models import User
 from onyx.db.persona import mark_persona_as_deleted
 from onyx.db.persona import upsert_persona
-from onyx.tools.built_in_tools import get_builtin_tool
+from onyx.db.tools import get_builtin_tool
 from onyx.tools.tool_implementations.search.search_tool import SearchTool
 from onyx.utils.errors import EERequiredError
 from onyx.utils.variable_functionality import (
