+++ backend/onyx/db/tools.py
@@ -1,5 +1,8 @@
 from typing import Any
 from typing import cast
+from typing import Type
+from typing import TYPE_CHECKING
+from typing import Union
 from uuid import UUID
 
 from sqlalchemy import select
@@ -9,6 +12,19 @@ from onyx.db.models import Tool
 from onyx.server.features.tool.models import Header
 from onyx.utils.headers import HeaderItemDict
 from onyx.utils.logger import setup_logger
+
+if TYPE_CHECKING:
+    from onyx.tools.tool_implementations.images.image_generation_tool import (
+        ImageGenerationTool,
+    )
+    from onyx.tools.tool_implementations.knowledge_graph.knowledge_graph_tool import (
+        KnowledgeGraphTool,
+    )
+    from onyx.tools.tool_implementations.search.search_tool import SearchTool
+    from onyx.tools.tool_implementations.web_search.web_search_tool import WebSearchTool
+    from onyx.tools.tool_implementations.okta_profile.okta_profile_tool import (
+        OktaProfileTool,
+    )
 
 logger = setup_logger()
 
@@ -104,3 +120,45 @@ def delete_tool__no_commit(tool_id: int, db_session: Session) -> None:
 
     db_session.delete(tool)
     db_session.flush()  # Don't commit yet, let caller decide when to commit
+
+
+def get_builtin_tool(
+    db_session: Session,
+    tool_type: Type[
+        Union[
+            "SearchTool",
+            "ImageGenerationTool",
+            "WebSearchTool",
+            "KnowledgeGraphTool",
+            "OktaProfileTool",
+        ]
+    ],
+) -> Tool:
+    """
+    Retrieves a built-in tool from the database based on the tool type.
+    """
+    # local import to avoid circular import. DB layer should not depend on tools layer.
+    from onyx.tools.built_in_tools import BUILT_IN_TOOL_MAP
+
+    tool_id = next(
+        (
+            in_code_tool_id
+            for in_code_tool_id, tool_cls in BUILT_IN_TOOL_MAP.items()
+            if tool_cls.__name__ == tool_type.__name__
+        ),
+        None,
+    )
+
+    if not tool_id:
+        raise RuntimeError(
+            f"Tool type {tool_type.__name__} not found in the BUILT_IN_TOOLS list."
+        )
+
+    db_tool = db_session.execute(
+        select(Tool).where(Tool.in_code_tool_id == tool_id)
+    ).scalar_one_or_none()
+
+    if not db_tool:
+        raise RuntimeError(f"Tool type {tool_type.__name__} not found in the database.")
+
+    return db_tool
